basic lambda for extracting numerical data
